+++ tools/garaga_rs/src/final_exp_witness/mod.rs
@@ -1,0 +1,2 @@
+pub mod bls12_381_final_exp_witness;
+pub mod bn254_final_exp_witness;
