+++ tools/garaga_rs/src/python_bindings/hades_permutation.rs
@@ -1,0 +1,48 @@
+use lambdaworks_crypto::hash::poseidon::{starknet::PoseidonCairoStark252, Poseidon};
+use lambdaworks_math::{
+    elliptic_curve::short_weierstrass::curves::{
+        bls12_381::field_extension::BLS12381PrimeField, bn_254::field_extension::BN254PrimeField,
+    },
+    field::{
+        element::FieldElement, fields::fft_friendly::stark_252_prime_field::Stark252PrimeField,
+        traits::IsPrimeField,
+    },
+    traits::ByteConversion,
+};
+use pyo3::{
+    types::{PyBytes, PyInt, PyList, PyTuple},
+    {prelude::*, wrap_pyfunction},
+};
+
+#[pyfunction]
+pub fn hades_permutation(
+    py: Python,
+    py_value_1: &Bound<'_, PyBytes>,
+    py_value_2: &Bound<'_, PyBytes>,
+    py_value_3: &Bound<'_, PyBytes>,
+) -> PyResult<PyObject> {
+    let byte_slice_1: &[u8] = py_value_1.as_bytes();
+    let byte_slice_2: &[u8] = py_value_2.as_bytes();
+    let byte_slice_3: &[u8] = py_value_3.as_bytes();
+
+    let mut state: Vec<FieldElement<Stark252PrimeField>> = vec![
+        FieldElement::<Stark252PrimeField>::from_bytes_be(byte_slice_1)
+            .expect("Unable to convert first param from bytes to FieldElement"),
+        FieldElement::<Stark252PrimeField>::from_bytes_be(byte_slice_2)
+            .expect("Unable to convert second param from bytes to FieldElement"),
+        FieldElement::<Stark252PrimeField>::from_bytes_be(byte_slice_3)
+            .expect("Unable to convert third param from bytes to FieldElement"),
+    ];
+
+    PoseidonCairoStark252::hades_permutation(&mut state);
+
+    let py_tuple = PyTuple::new_bound(
+        py,
+        state.iter().map(|fe| {
+            let fe_bytes = fe.to_bytes_be();
+            PyBytes::new_bound(py, &fe_bytes)
+        }),
+    );
+
+    Ok(py_tuple.into())
+}
